requirements.txt
APScheduler==3.9.1
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
DateTimeRange==1.2.0
idna==3.4
lxml==4.9.1
mbstrdecoder==1.1.1
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
typepy==1.3.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
